% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllComboFit.R
\name{AllPossibleCalSamples}
\alias{AllPossibleCalSamples}
\title{Determines All Calibration Sample Subsets}
\usage{
AllPossibleCalSamples(AllData, Controls = NULL, min = 8,
  max = nrow(Controls), Control.Range = 85:95)
}
\arguments{
\item{AllData}{A two-dimensional list seperated first by Jobs, then by Plates, containing the data for each experiment.}

\item{Controls}{data.frame containing the Control values for this experiment. Must have column names of Actual.Name, Name, Dosage; where Actual.Name is what is found exactly in the data sets, Name is the desired output name (often the same as Actual.Name), and Dosage is the Dosage for each control.}

\item{min}{The smallest subset of controls to use for the calibration sample.}

\item{max}{The largest subset of controls to use for the calibration sample. Usually this is the total number of controls.}

\item{Control.Range}{The row range describing which samples are part of the control list.}
}
\description{
This function finds all possible subsets from min to max size of the Control variables.
}
