% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallDosage.R
\name{BuildSummary}
\alias{BuildSummary}
\title{Gets the Predicted Dosage from Combined Dosage Calls}
\usage{
BuildSummary(Job, NotDone, PlateSz = 84, assays = c("Assay3", "Assay2", "Assay1",
  "Assay3#", "Assay1#"),
  dose.dir = "/path/to/data",
  Summary.File = paste0(dose.dir, "../summary/20", substr(Job, 2, 3),
  "_ProductionSummary.csv"), header = c("RecId", "Job", "#Samples", "Name",
  "Clone-plant_ID", "Entry", "TissID", "Clone ID", "Rack", "RackOrder",
  "Well Position", "DNA Well", "Pedigree", "Pedigree Long",
  c(matrix(sapply(assays[-grep("#", assays)], function(x) {     c(paste0("(", x,
  ")", c(" PCR", " RackOrder", " PCR Well", " A1 dRn", " A2 dRn")), paste0(x,
  c("", "_Pr", "_QC"))) }), nrow = 1)), c(matrix(sapply(assays[grep("#",
  assays)], function(x) c(paste0("(", x, ")", c(" PCR", " RackOrder",
  " PCR Well", " CT", " CT PEF", " dCT", " RQ")), paste0(x, c("", "_Pr",
  "_QC")))), nrow = 1), "Action", "uid")), num.cols = c("CT", "CT.PEF", "dCT",
  "RQ", "Allele1.Delta.Rn", "Allele2.Delta.Rn", matrix(rbind(assays,
  paste0(assays, "_Pr")), nrow = 1), paste0(assays, "_QC")),
  Assay.Cols = c("PCR", "RackOrder", "PCR Well", "A1 dRn", "A2 dRn", "CT",
  "CT PEF", "dCT", "RQ"),
  consol.cols.num = c(c(matrix(sapply(assays[-grep("#", assays)], function(x)
  {     c(paste0("(", x, ")", c(" A1 dRn", " A2 dRn")), paste0(x, c("", "_Pr")))
  }), nrow = 1)), c(matrix(sapply(assays[grep("#", assays)], function(x)
  c(paste0("(", x, ")", c(" CT", " CT PEF", " dCT", " RQ")), paste0(x, c("",
  "_Pr")))), nrow = 1))),
  consol.cols.str = c(c(matrix(sapply(assays[-grep("#", assays)], function(x)
  {     c(paste0("(", x, ")", c(" PCR", " RackOrder", " PCR Well"))) }), nrow =
  1)), c(matrix(sapply(assays[grep("#", assays)], function(x) c(paste0("(", x,
  ")", c(" PCR", " RackOrder", " PCR Well")))), nrow = 1))))
}
\arguments{
\item{Job}{Name of the current job.}

\item{NotDone}{Which PCR plates of the current job have yet to be added to the summary?}

\item{PlateSz}{Size of the DNA plates excluding controls. For example, a 96 well DNA plate with 12 controls would have PlateSz = 84. Used to index the output.}

\item{assays}{List of all possible assays. Default = c('Assay3','Assay2','Assay1','Assay3#','Assay1#')}

\item{dose.dir}{Directory of where to find the Dosage.csv files. Should have sub-directory structure of dose.dir/<Job>/<PCR>.Dosage.csv}

\item{Summary.File}{Name of the current summary file}
}
\description{
Checks the different models' accuracy against the controls and then sets the predicted values to be the best model.
}
\keyword{All}
\keyword{Combine,}
\keyword{Models,}
\keyword{Output,}
\keyword{Predicted}
