% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresenceAbsenceCaller.R
\name{CallAllPA}
\alias{CallAllPA}
\title{Runs CallPA() on all files in a directory}
\usage{
CallAllPA(PAData, sheetname, Controls, output.dir, outlier.threshold = 3, header = c("..."), console = FALSE)
}
\arguments{
\item{PAData}{A list of lists of each plate's Quant Studio Results file for Presence/Absence. Must contain Allele1.Delta.Rn and Allele2.Delta.Rn}

\item{sheetname}{Name of the sheet containing Allele Delta RN values}

\item{Controls}{File containing the known presence/absence information. In general this contains one column for each type of assay.}

\item{output.dir}{The directory to write the output files to, Plots will be placed in Plots sub directory}

\item{outlier.threshold}{The number of standard deviations away from the mean a sample must be to be considered an outlier}
}
\description{
This function determined the Presence/Absence of the appropriate assay based on the Quant Studio output file.
}
\keyword{Assay}
\keyword{Calling,}
\keyword{Dosage}
\keyword{Presence-Absence,}
