% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallDosage.R
\name{CallDosageAPI}
\alias{CallDosageAPI}
\title{Main Dosage Calling Function for the API}
\usage{
CallDosageAPI(Control.Dir, Control.File = NULL, Directory = NULL,
  Job = NULL, sheet = "Results", Skip.List = NULL, Console = F,
  Target.Column = "Target.Name", Target = "Assay1", Output.Dir = NULL,
  assays = c("Assay3", "Assay2", "Assay1", "Assay1#", "Assay3#"), RQ.Threshold = 7,
  Output.File = paste0(Output.Dir, "/", c(Job, basename(getwd()))[1],
  ".Dosage.csv"), REF.Vars = c("CT"), swap.file = paste0(Output.Dir,
  "/src/Swaps.csv"), header = c("..."))
}
\arguments{
\item{Control.Dir}{Directory containing the Control Files (Ignored, if Control.File is specified)}

\item{Control.File}{If specified, these controls will be used for each job}

\item{Directory}{Name of the directory storing the raw data for the jobs (Output from Quant Studio)}

\item{Job}{Name of the specific Job to run now during this call.}

\item{sheet}{Name of the sheet in the Raw Data where RQ, CT, and delta CT values are stored. Default is "Results" for an unchanged raw data file.}

\item{Skip.List}{List of Jobs to skip when searching for unupdated jobs. Should be a .csv file with a single column of Job names.}

\item{Console}{Should plots should be printed to console instead of output directory? (default = False)}

\item{Target.Column}{Column name that lists the target assay. For standard Quant Studio output, this is "Target.Name".}

\item{Output.Dir}{Name of the directory to store the files. The path should be relative to Directory's path.}

\item{assays}{List of possible assays. Default is  c('Assay3','Assay2','Assay1','Assay1#','Assay3#'), but this could simply match the target on the raw data file.}

\item{RQ.Threshold}{If RQ goes above this value we know something went wrong with the data. Default = 7, since RQ should never be above 5.}

\item{Output.File}{Name of output file containing each model's dosage/probability as well as the overall dosage/probability.}

\item{swap.file}{Location of file containing any sample changes during tissue sampling.}

\item{header}{All the relevant variables for the final output.}

\item{Ref.Vars}{Generate a new column containing the reference's value for any Ref.Vars. Default is CT to generate CT.PEF.}
}
\description{
Uses multiple models to fit the data, then uses each of these models to determine both the best fit, and liklihood of that fit. The models used in this algorithm are 
\describe{
\item{AllComboFit}{Following the \eqn{2^-\Delta\Delta Ct} method, this model considers all possible combinations of controls in order to generate a calibration \eqn{\Delta Ct} value. }
\item{Semi-Supervised K-means}{Uses standard unsupervised K-means clustering; however, by fixing controls values to their respective clusters it takes advantage of known information.}
\item{Linear Model}{Again following the \eqn{2^-\Delta\Delta Ct}, this method uses the fact that \eqn{log2(Dosage) ~ \Delta\Delta Ct}, which simplifies into \eqn{\log2(Dosage) ~ a + b\Delta Ct}. This way there is no dependence on finding a proper calibration sample, and we just use regression to determine \emph{a} and \emph{b}.}
\item{Simulated Manual Call}{Aligns the sorted RQ values and finds the natural clusters in the RQ values. It then labels this clusters from 0 to 4.}
}
}
\keyword{Calling,}
\keyword{Dosage}
\keyword{PCR,}
\keyword{Polyploid,}
\keyword{Real}
\keyword{Tetraploid,}
\keyword{Time}
\keyword{qPCR,}
