% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallDosage.R
\name{CallPA}
\alias{CallPA}
\title{Calls Presence/Absence for a given input file}
\usage{
CallPA(File, Control.File = "Controls.csv", Output.Dir = "Output",
  Directory = NULL, sheet = "Results", Console = F,
  Output.File = "PA.csv")
}
\arguments{
\item{File}{Name of the file containing the raw data from a Quant Studio qPCR run. The column Clone-plant_ID must be added.}

\item{Control.File}{Name of the file containing the controls. Must contain 2 columns, Name and Presence, with Name matching the values in Clone-Plant_ID for control samples. (Default = Controls.csv)}

\item{Directory}{The directory containing the necessary files. Default to NULL which will use current directory.}

\item{sheet}{The name in File where Clone-Plant_ID and RQ values are stored. Default = "Results".}

\item{Console}{Should plots be printed to the console? Default is False which prints the plots to Output.Dir.}

\item{Output.File}{The name of the output file containing the final dosage calls. Defaults to Dosage.csv.}
}
\value{
A dataset containing the dosage calls of each of the methods and a final determination of the dosage and likelihood.
}
\description{
Uses a clustering algorithm to cluster the 3 categories: Positive, Negative, NTC on the file provided.
Requires known control values as specified in the Control.File.
}
\examples{
# Obtain the appropriate Control and Raw Data files. You can find these in the
# data-raw/ folder of the Quadrophenia package files.They should match the data below.

# A 41 line header
# Followed by the output of QuantStudio's Results sheet
# Note that Clone-plant_ID must be added manually to this file

require(Quadrophenia)
RawData.Assay2[1:41,1] # Header
RawData.Assay2[42,]    # Columns of data

# The controls file has two columns: Clone-plant_ID and Presence. Clone-plant_ID must
# match the Clone-plant_ID columns for the associated controls in the raw data file.

head(ControlsAssay2)

# Then run CallPA() with these files as inputs

output = CallPA(File='RawData.Assay2.xls',Control.File='ControlsAssay2.csv')

# The output directory Output/ should contain 1 .csv file and a Plots/ directory.
# The Plots/ directory contains one plot showing the clustering of Positive, Negative,
# and Undetermined calls for the expression of the target & reference assays.

head(output)

## Other ways to run CallPA()
# You could also specifiy an output file or a directory to look for the raw data files

dir.create('Files/')
file.copy(c('RawData.Assay2.xls','ControlsAssay2.csv'),'Files/')
output = CallPA(File='RawData.Assay2.xls',
                Control.File='ControlsAssay2.csv',
                Directory='Files/',Output.Dir='NewOutput')

# Or specify the files in Files/ wihout specifying the directory. You can
# also change the name of the output file to something more informative.

output = CallPA(File='Files/RawData.Assay2.xls',
                Control.File='Files/ControlsAssay2.csv',
                Output.Dir='ThirdOutput',
                Output.File='ThirdOutputPA.csv')

# Duplicates will automatically be renamed to avoid conflicts

output = CallPA(File='Files/RawData.Assay2.xls',
                Control.File='Files/ControlsAssay2.csv',
                Output.Dir='ThirdOutput',
                Output.File='ThirdOutputPA.csv')

# Clean up the test files
unlink(c('Files','Output','ThirdOutput'),recursive=T)
}
\seealso{
\code{\link{CallDosage}} for calling the dosage of provided data
}
