% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresenceAbsenceCaller.R
\name{CallSinglePA}
\alias{CallSinglePA}
\title{Determines if a Given Assay is Present in a Sample}
\usage{
CallSinglePA(df, Plate, sheetname, Controls, output.dir,
  outlier.threshold = 2, console = FALSE)
}
\arguments{
\item{sheetname}{Name of the sheet containing Allele Delta RN values}

\item{Controls}{File containing the known presence/absence information. In general this contains one column for each type of assay.}

\item{output.dir}{The directory to write the output files to, Plots will be placed in Plots sub directory}

\item{outlier.threshold}{The number of standard deviations away from the mean a sample must be to be considered an outlier}
}
\description{
This function call the semi-supervised clustering algorithm for each plate/job of a given dataset.
}
\keyword{Assay}
\keyword{Calling,}
\keyword{Dosage}
\keyword{Presence-Absence,}
