% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervisedClustering.R
\name{ChooseClosestCluster}
\alias{ChooseClosestCluster}
\title{Determine Closest Cluster to Sample}
\usage{
ChooseClosestCluster(ClustMeans, val, nClust)
}
\arguments{
\item{ClustMeans}{An \emph{nClust} x \emph{p} numeric Matrix of the mean values of each paramater for each cluster.}

\item{val}{The specific sample for which the closest cluster is desired.}

\item{nClust}{Expected total number of clusters (i.e the possible dosage values). Default is 5 for tetraploids.}
}
\description{
Helper function for MyClustering which chooses the closest cluster for a given sample
}
