% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallDosage.R
\name{CombineDosages}
\alias{CombineDosages}
\title{Combine Each of the Models}
\usage{
CombineDosages(Dosages, Manual, SemiSuper, Target, Jobs = names(Dirs),
  Plates = lapply(lapply(seq(length(Dirs)), function(x) {    
  unlist(gsub(paste0(".xls|.xlsx|.txt|.csv", paste0(names(Dirs)[x], "-")), "",
  Dirs[[x]])) }), as.character))
}
\arguments{
\item{Dosages}{The output from \code{\link{GetDosages}} containing the SVM, Multinomial, and QS dosage predictions.}

\item{Manual}{The Manual calling simulated model's output. See \code{\link{ManualCall.Trials}} for more information}

\item{SemiSuper}{The semi-supervised clustering model's output. See \code{\link{MyClustering}} for more information}

\item{Jobs}{The names of each of the folders for each job.}

\item{Plates}{Names of the individual excel datasets within each Job. Usually <Plate Name>.xlsx}
}
\description{
Combine each of the different models into a single dataframe. You could add more models and combine them here as desired.
}
\keyword{All}
\keyword{Combine,}
\keyword{Models}
\keyword{Output,}
