% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllComboFit.R
\name{ComputeScores}
\alias{ComputeScores}
\title{Measures the Accuracy of a Model vs. Control Dosages}
\usage{
ComputeScores(AllDosages, Controls = NULL)
}
\arguments{
\item{AllDosages}{A serial data frame of all the relevant values including Job, Plate, RQ, CT, delta CT, and Sample Name.}

\item{Controls}{data.frame containing the Control values for this experiment. Must have column names of Actual.Name, Name, Dosage; where Actual.Name is what is found exactly in the data sets, Name is the desired output name (often the same as Actual.Name), and Dosage is the Dosage for each control.}
}
\description{
This measures the percentage of controls that agree with the predicted values.
}
