% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManualDosageCalling.R
\name{CountSections}
\alias{CountSections}
\title{Helper Function for Manual Call}
\usage{
CountSections(RQ.Vals, by.val = 0.1, min = 0, max = 5)
}
\arguments{
\item{RQ.Vals}{A list of RQ values. Runs faster if this is already sorted.}

\item{by.val}{The steps between from and to.}

\item{min}{The smallest RQ value to be considered.}

\item{max}{The largest RQ value to be considered. NULL implies no limit.}
}
\description{
This function counts the the number of RQ values in a given range.
}
