% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervisedClustering.R
\name{distance}
\alias{distance}
\title{Distance Measure}
\usage{
distance(x, y)
}
\arguments{
\item{x}{A numeric vector of \emph{p} numbers.}

\item{y}{A numeric vector of \emph{p} numbers.}
}
\description{
The Euclidean metric for measuring the \emph{p} dimensional distance between samples.
}
