% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllComboFit.R
\name{DosageOfAllCombos}
\alias{DosageOfAllCombos}
\title{Get the Dosage for All Combinations of Controls}
\usage{
DosageOfAllCombos(AllData, all.control.combos = NULL, Results.Only = FALSE,
  Controls = NULL, min = 1, max = nrow(Controls))
}
\arguments{
\item{AllData}{A two-dimensional list seperated first by Jobs, then by Plates, containing the data for each experiment.}

\item{all.control.combos}{All possible combinations of the control variables.}

\item{Results.Only}{FALSE means that the AllData is actually a three-dimensonal list with the third dimension being each sheet in the data sets. If AllData only contains the "Results' section then set this to TRUE.}

\item{Controls}{data.frame containing the Control values for this experiment. Must have column names of Actual.Name, Name, Dosage; where Actual.Name is what is found exactly in the data sets, Name is the desired output name (often the same as Actual.Name), and Dosage is the Dosage for each control.}

\item{min}{The smallest subset of controls to use for the calibration sample.}

\item{max}{The largest subset of controls to use for the calibration sample. Usually this is the total number of controls.}
}
\description{
Uses all possible subsets of control values in order to generate a calibration delta CT value.
}
