% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllComboFit.R
\name{DosageOfGivenCombo}
\alias{DosageOfGivenCombo}
\title{Find the Dosage Given a Subset of Controls}
\usage{
DosageOfGivenCombo(AllData, Controls, Results.Only = FALSE,
  Target = "Assay1 CN ", Reference = "PEF CN", Num.Iters = 5)
}
\arguments{
\item{AllData}{A two-dimensional list seperated first by Jobs, then by Plates, containing the data for each experiment.}

\item{Controls}{data.frame containing the Control values for this experiment. Must have column names of Actual.Name, Name, Dosage; where Actual.Name is what is found exactly in the data sets, Name is the desired output name (often the same as Actual.Name), and Dosage is the Dosage for each control.}

\item{Results.Only}{FALSE means that the AllData is actually a three-dimensonal list with the third dimension being each sheet in the data sets. If AllData only contains the "Results' section then set this to TRUE.}

\item{Target}{The reference gene for calling the dosage. Must follow the exact name in datasets.}

\item{Num.Iters}{The number of iterations to run on the dosage calling algorithm. Each iteration refines the "by" variable by 10. See ManualCall.Trials for more information.}

\item{Target}{The target gene for calling the dosage. Must follow the exact name in datasets.}
}
\description{
This function determines the calibration sample based on the provided subset of controls. It then computes the RQ values with that samples delta CT value and finds the dosage from those.
}
