% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervisedClustering.R
\name{GetConf}
\alias{GetConf}
\title{Measures the Confidence in Prediction}
\usage{
GetConf(fitAllClust, limit, index = 1:2, lbl.rng = 0:2)
}
\arguments{
\item{fitAllClust}{A 2 dimensional list containing a data frame (named \emph{df}), with the paramater values for each sample along with the predicted cluster, and  the ClustMeans matrix, an \emph{nClust} x \emph{p} numeric Matrix of the mean values of each paramater for each cluster.}

\item{limit}{The number of standard deviations from the mean that would warrant being called an outlier.}

\item{index}{Which parameters to be used for measuring confidence. Default assumes only 2 features, but often this number is 1:\emph{p}.}
}
\description{
Used to measure the confidence of each dosage call following standard multivariate normal confidence intervals.
}
