% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModel.R
\name{GetDosageForSinglePlate}
\alias{GetDosageForSinglePlate}
\title{Linear Model on a Single Plate}
\usage{
GetDosageForSinglePlate(Plate, job = NA, plate = NA, Controls = NULL,
  MeanRQ = NULL, Target = "Assay1")
}
\arguments{
\item{Plate}{A serial data frame of all the relevant values including Job, Plate, RQ, CT, delta CT, and Sample Name for a single plate.}

\item{job}{Name of the job that this plate is under.}

\item{plate}{Name of the plate.}

\item{Controls}{data.frame containing the Control values for this experiment. Must have column names of Actual.Name, Name, Dosage; where Actual.Name is what is found exactly in the data sets, Name is the desired output name (often the same as Actual.Name), and Dosage is the Dosage for each control.}

\item{MeanRQ}{The Mean RQ value across all Jobs/Plates. Used for scaling the data to provide a more balanced dosage call. Default=NULL means no scaling.}
}
\description{
This function performs a linear fit of log2(Dosage) as a function of delta CT for a single plate.
It also computes a Multinomial distribution of Dosage ~ f(P(RQ)) where P is a 2nd degree polynomial.
}
