% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManualDosageCalling.R
\name{GetManualCall}
\alias{GetManualCall}
\title{Performs a Simulated Manual Call}
\usage{
GetManualCall(AllData, Target = "Assay1")
}
\arguments{
\item{AllData}{A serial data frame of all the relevant values including Job, Plate, RQ, CT, delta CT, and Sample Name.}

\item{Target}{The target gene for calling the dosage. Must follow the exact name in datasets.}
}
\description{
Following the pre-existing manual method of determining dosage from RQ values, this algorithm finds the obvious clusters among the RQ values and assigns each cluster a dosage.
}
