% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallDosage.R
\name{GetPredicted}
\alias{GetPredicted}
\title{Gets the Predicted Dosage from Combined Dosage Calls}
\usage{
GetPredicted(df, Controls, Output.Dir, Target = "Assay1",
  Methods = c("Linear", "Multinomial", "Manual", "SemiSupervised",
  "QS.Dosage"), Console = F, LL.min = 0.5, nulliplex.thresh = 0.02)
}
\arguments{
\item{df}{The dataset built from joining the results of each method specified in methods. Use \code{\link{CombineDosages}}}

\item{Controls}{The Controls with Well information included}

\item{Output.Dir}{The directory to print the resulting plots to. Uneccesary if \code{Console=T}.}

\item{Target}{The target assay. This is used primarily to rename the "Predicted" and "Probability" columns to the correct names. The assay names Assay1,Assay1#,Assay1CN, Assay1 CN, etc. will all generate a column named Assay1# and Assay1#_Pr.}

\item{Methods}{The different methods combined with \code{\link{CombineDosages}}. Must match \code{df}'s column names for these predictions.}

\item{Console}{Should the plots be printed to the console and not saved to \code{Output.Dir}? (Default=F).}

\item{LL.min}{Any predictions with likelihood less than this will be ignored. Predictions with likelihood greater than this will be included in the final predictions weighted average. (Default=0.5)}

\item{nulliplex.thres}{Final predictions with RQ < \code{nulliplex.thresh} will be set to nulliplex. Necessary the reaction fails for all nulliplex controls (very probable.). (Default = 0.2)}

\item{Controls}{the list of Controls for the dataset \code{df}.}
}
\description{
Checks the different models' accuracy against the controls and then sets the predicted values to be the best model.
}
\keyword{All}
\keyword{Combine,}
\keyword{Models,}
\keyword{Output,}
\keyword{Predicted}
