% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallDosage.R
\name{GetSemiSuperClusters}
\alias{GetSemiSuperClusters}
\title{Driver Function for Calling Semi-Supervised Clustering Algorithm}
\usage{
GetSemiSuperClusters(AllData, Target = "Assay1", Controls = NULL)
}
\arguments{
\item{AllData}{data.frame containing the relevent data. There must be a column for each of the VarNames/VarIndices, and the control samples must exactly match the Contols data.frame.}

\item{Target}{The target gene for calling the dosage. Must follow the exact name in datasets.}

\item{Controls}{data.frame containing the Control values for this experiment. Must have column names of Actual.Name, Name, Dosage; where Actual.Name is what is found exactly in the data sets, Name is the desired output name (often the same as Actual.Name), and Dosage is the Dosage for each control. If NULL then this algorithm is no different than K-Means clustering, and will not return correct dosage numbers..}

\item{VarIndices}{Index of the columns that contain variables to be used in the clustering process.}

\item{VarNames}{Names of the columns that contain variables to be used in the clustering process. Using this will replace VarIndices with the indices of the provided VarNames.}
}
\description{
This function call the semi-supervised clustering algorithm for each plate/job of a given dataset.
}
\keyword{K-Means}
\keyword{Semi-Supervised,}
