% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallDosage.R
\name{LoadAllData}
\alias{LoadAllData}
\title{Loads All Data Sets}
\usage{
LoadAllData(Dirs = NULL, sheet = NULL, want.list = FALSE,
  strings.factor = FALSE)
}
\description{
Helper Function for Call Dosage
}
