% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallDosage.R
\name{MakeSwaps}
\alias{MakeSwaps}
\title{Swaps any changed sample values}
\usage{
MakeSwaps(df, swap.dir, Job)
}
\arguments{
\item{df}{The dataframe containing each of the models' values produced from CombineDosageCalls/CombinedPACalls}

\item{swap.dir}{The directory containing Swaps.csv}

\item{Job}{The name of the current job being called}
}
\description{
Reads the Swaps.csv file in src and applies any substitions found
}
\keyword{Sampling}
\keyword{Substitute,}
\keyword{Swap,}
\keyword{Swaps.csv,}
\keyword{Tissue,}
