% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManualDosageCalling.R
\name{ManualCall}
\alias{ManualCall}
\title{Simulated Manual Dosage Call}
\usage{
ManualCall(RQ.Vals, by.val = 0.1, min = 0, max = NULL,
  bin.threshold = 0.01, min.threshold = 1)
}
\arguments{
\item{RQ.Vals}{A list of RQ values. Runs faster if this is already sorted.}

\item{by.val}{The steps between from and to.}

\item{min}{The smallest RQ value to be considered.}

\item{max}{The largest RQ value to be considered. NULL implies no limit.}

\item{bin.threshold}{The minimum percentage that a bin (cluster) is allowed to be relative to all the values.}

\item{min.threshold}{Minimum number of samples allowed in each cluster.}
}
\description{
Simulates pre-existing manual calling method by clustering sorted RQ values.
}
