% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManualDosageCalling.R
\name{ManualCall.Trials}
\alias{ManualCall.Trials}
\title{Runs Multiple Iterations of Manual Call}
\usage{
ManualCall.Trials(RQ.Vals, ploidy = 4, from = 0.01, to = 1, by = 0.01,
  min = 0, max = NULL, bin.threshold = 0.01)
}
\arguments{
\item{RQ.Vals}{A list of RQ values. Runs faster if this is already sorted.}

\item{ploidy}{The number of sets of chromosomes in a cell.}

\item{from}{The smallest interval length for calling Manual Call with.}

\item{to}{The largest interval length for calling Manual Call with.}

\item{by}{The steps between from and to.}

\item{min}{The smallest RQ value to be considered.}

\item{max}{The largest RQ value to be considered. NULL implies no limit.}

\item{bin.threshold}{The minimum percentage that a bin (cluster) is allowed to be relative to all the values.}
}
\description{
This function runs multiple iterations of manual call with smaller and smaller refinement of clustering.
}
