% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllComboFit.R
\name{MeasureConfidence}
\alias{MeasureConfidence}
\title{Measures the Confidence of An AllComboFit Output}
\usage{
MeasureConfidence(df)
}
\arguments{
\item{df}{An AllComboFit output from DosageOfAllCombos}
}
\description{
Generates the probability of each dosage call by comparing the mean and standard deviation of all fits from an AllComboFit output.
}
