% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresenceAbsenceCaller.R
\name{ParsePAComments}
\alias{ParsePAComments}
\title{Determines if a Given Assay is Present in a Sample}
\usage{
ParsePAComments(df)
}
\arguments{
\item{sheetname}{Name of the sheet containing Allele Delta RN values}

\item{control.file}{Location of the file containing the known Presence Absence values for each assay}

\item{output.dir}{The directory to write the output files to, Plots will be placed in Plots sub directory}

\item{outlier.threshold}{The number of standard deviations away from the mean a sample must be to be considered an outlier}
}
\description{
This function call the semi-supervised clustering algorithm for each plate/job of a given dataset.
}
\keyword{Assay}
\keyword{Calling,}
\keyword{Dosage}
\keyword{Presence-Absence,}
