% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresenceAbsenceCaller.R
\name{PlotPA}
\alias{PlotPA}
\title{Print Output for Presence/Absence Calling}
\usage{
PlotPA(df, output.dir = "", Plate = "", Assay = "", console = FALSE,
  badPlot = F)
}
\description{
Helper Function for CallSinglePA
}
