% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallDosage.R
\name{RemoveDuplicateNames}
\alias{RemoveDuplicateNames}
\title{Removes Duplicated Names in List}
\usage{
RemoveDuplicateNames(data, Dirs, Controls = NULL, Jobs = names(Dirs),
  Plates = lapply(lapply(seq(length(Dirs)), function(x) {    
  unlist(gsub(paste0(".xls|.xlsx|.txt|.csv|", paste0(names(Dirs)[x], "-")), "",
  Dirs[[x]])) }), as.character), SampleIndex = which(names(data[[1]][[1]]) ==
  "Sample.Name"), Target.Name = "Assay1 CN ")
}
\description{
Helper Function for Call Dosage
}
